// src/modules/chatbot/domain/enums/intent-name.enum.ts
export enum IntentName {
  GREETING = 'greeting',
  GOODBYE = 'goodbye',
  THANKS = 'thanks',
  
  // University-related
  ABOUT_UNIVERSITIES = 'about_universities',
  LIST_UNIVERSITIES = 'list_universities',
  UNIVERSITY_INFO = 'university_info',
  UNIVERSITY_LOCATION = 'university_location',
  
  // Programs
  LIST_PROGRAMS = 'list_programs',
  PROGRAM_INFO = 'program_info',
  PROGRAM_CAREERS = 'program_careers',
  
  // Admission
  ADMISSION_REQUIREMENTS = 'admission_requirements',
  HOW_TO_APPLY = 'how_to_apply',
  REQUIRED_DOCUMENTS = 'required_documents',
  
  // Fees
  TUITION_FEES = 'tuition_fees',
  SCHOLARSHIPS = 'scholarships',
  
  // Contact
  CONTACT_INFO = 'contact_info',
  WORKING_HOURS = 'working_hours',
  
  // Default
  UNKNOWN = 'unknown',
  HELP = 'help',
}